RugbyAnalyticsSolution/RugbyAnalyticsProject/requirements.txt
Django>=3.2.14
djangorestframework>=3.12.4
beautifulsoup4>=4.10.0
requests>=2.26.0
django-cors-headers>=3.7.0
pytz
sqlparse
pandas>=1.3.0
matplotlib>=3.4.3

